Separating filter parsing logic
