+++ src/Microsoft.SqlTools.ServiceLayer/ObjectExplorer/ObjectExplorerUtils.cs
@@ -8,6 +8,8 @@
 using System;
 using System.Threading;
 using Microsoft.SqlTools.ServiceLayer.ObjectExplorer.Nodes;
+using Microsoft.SqlTools.ServiceLayer.ObjectExplorer.Contracts;
+using System.Collections.Generic;
 
 namespace Microsoft.SqlTools.ServiceLayer.ObjectExplorer
 {
@@ -53,7 +55,7 @@ namespace Microsoft.SqlTools.ServiceLayer.ObjectExplorer
         /// <returns>A Tree Node that matches the condition, or null if no matching node could be found</returns>
         public static TreeNode? FindNode(TreeNode node, Predicate<TreeNode> condition, Predicate<TreeNode> filter, bool expandIfNeeded = false)
         {
-            if(node == null)
+            if (node == null)
             {
                 return null;
             }
@@ -76,5 +78,97 @@ namespace Microsoft.SqlTools.ServiceLayer.ObjectExplorer
             }
             return null;
         }
+
+        public static INodeFilter ConvertExpandNodeFilterToNodeFilter(NodeFilter filter, NodeFilterProperty filterProperty)
+        {
+            Type type = typeof(string);
+
+            var IsDateTime = filterProperty.Type == NodeFilterPropertyDataType.Date;
+
+            FilterType filterType = FilterType.EQUALS;
+            Boolean isNotFilter = false;
+
+            object filterValue = null;
+
+            switch (filterProperty.Type)
+            {
+                case NodeFilterPropertyDataType.String:
+                case NodeFilterPropertyDataType.Date:
+                case NodeFilterPropertyDataType.Choice:
+                    type = typeof(string);
+                    filterValue = filter.Value.ToString();
+                    break;
+                case NodeFilterPropertyDataType.Number:
+                    type = typeof(int);
+                    filterValue = filter.Value.ToObject<int>();
+                    break;
+                case NodeFilterPropertyDataType.Boolean:
+                    type = typeof(bool);
+                    filterValue = filter.Value.ToObject<bool>() ? 1 : 0;
+                    break;
+            }
+
+            switch (filter.Operator)
+            {
+                case NodeFilterOperator.Equals:
+                    filterType = FilterType.EQUALS;
+                    break;
+                case NodeFilterOperator.NotEquals:
+                    filterType = FilterType.EQUALS;
+                    isNotFilter = true;
+                    break;
+                case NodeFilterOperator.LessThan:
+                    filterType = FilterType.LESSTHAN;
+                    break;
+                case NodeFilterOperator.LessThanOrEquals:
+                    filterType = FilterType.LESSTHANOREQUAL;
+                    break;
+                case NodeFilterOperator.GreaterThan:
+                    filterType = FilterType.GREATERTHAN;
+                    break;
+                case NodeFilterOperator.GreaterThanOrEquals:
+                    filterType = FilterType.GREATERTHANOREQUAL;
+                    break;
+                case NodeFilterOperator.Contains:
+                    filterType = FilterType.CONTAINS;
+                    break;
+                case NodeFilterOperator.NotContains:
+                    filterType = FilterType.CONTAINS;
+                    isNotFilter = true;
+                    break;
+                case NodeFilterOperator.Between:
+                    filterType = FilterType.BETWEEN;
+                    break;
+                case NodeFilterOperator.NotBetween:
+                    filterType = FilterType.NOTBETWEEN;
+                    isNotFilter = true;
+                    break;
+            }
+
+
+            if (filter.Operator == NodeFilterOperator.Between || filter.Operator == NodeFilterOperator.NotBetween)
+            {
+                if (filterProperty.Type == NodeFilterPropertyDataType.Number)
+                {
+                    filterValue = filter.Value.ToObject<int[]>();
+                }
+                else if (filterProperty.Type == NodeFilterPropertyDataType.Date)
+                {
+                    filterValue = filter.Value.ToObject<string[]>();
+                }
+            }
+
+            return new NodePropertyFilter
+            {
+                Property = filterProperty.Name,
+                Type = type,
+                Values = new List<object> { filterValue },
+                IsNotFilter = isNotFilter,
+                FilterType = filterType,
+                IsDateTime = IsDateTime
+            };
+        }
+
+
     }
 }
