+++ src/Microsoft.SqlTools.ServiceLayer/ObjectExplorer/SmoModel/SmoChildFactoryBase.cs
@@ -148,92 +148,7 @@ namespace Microsoft.SqlTools.ServiceLayer.ObjectExplorer.SmoModel
                 foreach (var f in appliedFilters)
                 {
                     NodeFilterProperty filterProperty = filterDefinitions.FirstOrDefault(x => x.DisplayName == f.DisplayName);
-                    Type type = typeof(string);
-
-                    var IsDateTime = filterProperty.Type == NodeFilterPropertyDataType.Date;
-
-                    FilterType filterType = FilterType.EQUALS;
-                    Boolean isNotFilter = false;
-
-                    object filterValue = null;
-
-                    switch (filterProperty.Type)
-                    {
-                        case NodeFilterPropertyDataType.String:
-                        case NodeFilterPropertyDataType.Date:
-                        case NodeFilterPropertyDataType.Choice:
-                            type = typeof(string);
-                            filterValue = f.Value.ToString();
-                            break;
-                        case NodeFilterPropertyDataType.Number:
-                            type = typeof(int);
-                            filterValue = f.Value.ToObject<int>();
-                            break;
-                        case NodeFilterPropertyDataType.Boolean:
-                            type = typeof(bool);
-                            filterValue = f.Value.ToObject<bool>() ? 1 : 0;
-                            break;
-                    }
-
-                    switch (f.Operator)
-                    {
-                        case NodeFilterOperator.Equals:
-                            filterType = FilterType.EQUALS;
-                            break;
-                        case NodeFilterOperator.NotEquals:
-                            filterType = FilterType.EQUALS;
-                            isNotFilter = true;
-                            break;
-                        case NodeFilterOperator.LessThan:
-                            filterType = FilterType.LESSTHAN;
-                            break;
-                        case NodeFilterOperator.LessThanOrEquals:
-                            filterType = FilterType.LESSTHANOREQUAL;
-                            break;
-                        case NodeFilterOperator.GreaterThan:
-                            filterType = FilterType.GREATERTHAN;
-                            break;
-                        case NodeFilterOperator.GreaterThanOrEquals:
-                            filterType = FilterType.GREATERTHANOREQUAL;
-                            break;
-                        case NodeFilterOperator.Contains:
-                            filterType = FilterType.CONTAINS;
-                            break;
-                        case NodeFilterOperator.NotContains:
-                            filterType = FilterType.CONTAINS;
-                            isNotFilter = true;
-                            break;
-                        case NodeFilterOperator.Between:
-                            filterType = FilterType.BETWEEN;
-                            break;
-                        case NodeFilterOperator.NotBetween:
-                            filterType = FilterType.NOTBETWEEN;
-                            isNotFilter = true;
-                            break;
-                    }
-
-
-                    if (f.Operator == NodeFilterOperator.Between || f.Operator == NodeFilterOperator.NotBetween)
-                    {
-                        if (filterProperty.Type == NodeFilterPropertyDataType.Number)
-                        {
-                            filterValue = f.Value.ToObject<int[]>();
-                        }
-                        else if (filterProperty.Type == NodeFilterPropertyDataType.Date)
-                        {
-                            filterValue = f.Value.ToObject<string[]>();
-                        }
-                    }
-
-                    filters.Add(new NodePropertyFilter
-                    {
-                        Property = filterProperty.Name,
-                        Type = type,
-                        Values = new List<object> { filterValue },
-                        IsNotFilter = isNotFilter,
-                        FilterType = filterType,
-                        IsDateTime = IsDateTime
-                    });
+                    filters.Add(ObjectExplorerUtils.ConvertExpandNodeFilterToNodeFilter(f, filterProperty));
                 }
             }
